Create dummy
CUSTOMER CARE REGISTRY

TECHNOLOGY : CLOUD APP DEVELOPMENT
TEAM ID : PNT2022MID14874

PROJECT DESCRIPTION
This Application has been developed to help the customer in processing their complaints. The customers can raise the ticket with a detailed description of the issue. An Agent will be assigned to the Customer to solve the problem. Whenever the agent is assigned to a customer they will be notified with an email alert. Customers can view the status of the ticket till the service is provided.

ADMIN : The main role and responsibility of the admin are to take care of the whole process. Starting from Admin login followed by the agent creation and assigning the customer's complaints. Finally, He will be able to track the work assigned to the agent and a notification will be sent to the customer.

USER : They can register for an account. After the login, they can create the complaint with a description of the problem they are facing. Each user will be assigned with an agent. They can view the status of their complaint.

SOFTWARE REQUIRED
PYTHON
FLASK
DOCKER 

 TEAM MEMBERS
FARHEENSULTANA- Team Lead
CHANDI 
CHARULATHA 
AASHA  

ASSIGNMENT FINISHED
ASSIGNMENT 1
 ASSIGNMENT 2
 ASSIGNMENT 3
ASSIGNMENT 4